requirements.txt
discord.py>=2.3.0
google-genai>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
groq>=1.0.0
requests>=2.31.0
aiohttp
discord.py
google-genai
groq
Pillow
requests
sift-stack-py
